chore: format contracts using forge formatter

+++ contracts/ERC721Creator.sol
@@ -14,13 +14,20 @@ import "./token/ERC721/ERC721Base.sol";
 contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     using EnumerableSet for EnumerableSet.AddressSet;
 
-    constructor (string memory _name, string memory _symbol) ERC721Base(_name, _symbol) {}
+    constructor(string memory _name, string memory _symbol) ERC721Base(_name, _symbol) {}
 
     /**
      * @dev See {IERC165-supportsInterface}.
      */
-    function supportsInterface(bytes4 interfaceId) public view virtual override(ERC721Core, ERC721CreatorCore, AdminControl) returns (bool) {
-        return ERC721CreatorCore.supportsInterface(interfaceId) || ERC721Core.supportsInterface(interfaceId) || AdminControl.supportsInterface(interfaceId);
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(ERC721Core, ERC721CreatorCore, AdminControl)
+        returns (bool)
+    {
+        return ERC721CreatorCore.supportsInterface(interfaceId) || ERC721Core.supportsInterface(interfaceId)
+            || AdminControl.supportsInterface(interfaceId);
     }
 
     function _beforeTokenTransfer(address from, address to, uint256 tokenId, uint96 data) internal virtual override {
@@ -38,11 +45,14 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     /**
      * @dev See {ICreatorCore-registerExtension}.
      */
-    function registerExtension(address extension, string calldata baseURI, bool baseURIIdentical) external override adminRequired {
+    function registerExtension(address extension, string calldata baseURI, bool baseURIIdentical)
+        external
+        override
+        adminRequired
+    {
         requireNonBlacklist(extension);
         _registerExtension(extension, baseURI, baseURIIdentical);
     }
-
 
     /**
      * @dev See {ICreatorCore-unregisterExtension}.
@@ -96,9 +106,11 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     function setTokenURIExtension(uint256[] calldata tokenIds, string[] calldata uris) external override {
         requireExtension();
         require(tokenIds.length == uris.length, "Invalid input");
-        for (uint i; i < tokenIds.length;) {
+        for (uint256 i; i < tokenIds.length;) {
             _setTokenURIExtension(tokenIds[i], uris[i]);
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
     }
 
@@ -128,9 +140,11 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
      */
     function setTokenURI(uint256[] calldata tokenIds, string[] calldata uris) external override adminRequired {
         require(tokenIds.length == uris.length, "Invalid input");
-        for (uint i; i < tokenIds.length;) {
+        for (uint256 i; i < tokenIds.length;) {
             _setTokenURI(tokenIds[i], uris[i]);
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
     }
 
@@ -144,49 +158,74 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     /**
      * @dev See {IERC721CreatorCore-mintBase}.
      */
-    function mintBase(address to) public virtual override nonReentrant adminRequired returns(uint256) {
+    function mintBase(address to) public virtual override nonReentrant adminRequired returns (uint256) {
         return _mintBase(to, "", 0);
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintBase}.
      */
-    function mintBase(address to, string calldata uri) public virtual override nonReentrant adminRequired returns(uint256) {
+    function mintBase(address to, string calldata uri)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+        returns (uint256)
+    {
         return _mintBase(to, uri, 0);
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintBaseBatch}.
      */
-    function mintBaseBatch(address to, uint16 count) public virtual override nonReentrant adminRequired returns(uint256[] memory tokenIds) {
+    function mintBaseBatch(address to, uint16 count)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+        returns (uint256[] memory tokenIds)
+    {
         tokenIds = new uint256[](count);
-        uint256 firstTokenId = _tokenCount+1;
+        uint256 firstTokenId = _tokenCount + 1;
         _tokenCount += count;
 
-        for (uint i; i < count;) {
-            tokenIds[i] = _mintBase(to, "", firstTokenId+i);
-            unchecked { ++i; }
+        for (uint256 i; i < count;) {
+            tokenIds[i] = _mintBase(to, "", firstTokenId + i);
+            unchecked {
+                ++i;
+            }
         }
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintBaseBatch}.
      */
-    function mintBaseBatch(address to, string[] calldata uris) public virtual override nonReentrant adminRequired returns(uint256[] memory tokenIds) {
+    function mintBaseBatch(address to, string[] calldata uris)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+        returns (uint256[] memory tokenIds)
+    {
         tokenIds = new uint256[](uris.length);
-        uint256 firstTokenId = _tokenCount+1;
+        uint256 firstTokenId = _tokenCount + 1;
         _tokenCount += uris.length;
 
-        for (uint i; i < uris.length;) {
-            tokenIds[i] = _mintBase(to, uris[i], firstTokenId+i);
-            unchecked { ++i; }
+        for (uint256 i; i < uris.length;) {
+            tokenIds[i] = _mintBase(to, uris[i], firstTokenId + i);
+            unchecked {
+                ++i;
+            }
         }
     }
 
     /**
      * @dev Mint token with no extension
      */
-    function _mintBase(address to, string memory uri, uint256 tokenId) internal virtual returns(uint256) {
+    function _mintBase(address to, string memory uri, uint256 tokenId) internal virtual returns (uint256) {
         if (tokenId == 0) {
             ++_tokenCount;
             tokenId = _tokenCount;
@@ -207,7 +246,7 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     /**
      * @dev See {IERC721CreatorCore-mintExtension}.
      */
-    function mintExtension(address to) public virtual override nonReentrant returns(uint256) {
+    function mintExtension(address to) public virtual override nonReentrant returns (uint256) {
         requireExtension();
         return _mintExtension(to, "", 0, 0);
     }
@@ -215,7 +254,7 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     /**
      * @dev See {IERC721CreatorCore-mintExtension}.
      */
-    function mintExtension(address to, string calldata uri) public virtual override nonReentrant returns(uint256) {
+    function mintExtension(address to, string calldata uri) public virtual override nonReentrant returns (uint256) {
         requireExtension();
         return _mintExtension(to, uri, 0, 0);
     }
@@ -223,7 +262,7 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     /**
      * @dev See {IERC721CreatorCore-mintExtension}.
      */
-    function mintExtension(address to, uint80 data) public virtual override nonReentrant returns(uint256) {
+    function mintExtension(address to, uint80 data) public virtual override nonReentrant returns (uint256) {
         requireExtension();
         return _mintExtension(to, "", data, 0);
     }
@@ -231,52 +270,80 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     /**
      * @dev See {IERC721CreatorCore-mintExtensionBatch}.
      */
-    function mintExtensionBatch(address to, uint16 count) public virtual override nonReentrant returns(uint256[] memory tokenIds) {
+    function mintExtensionBatch(address to, uint16 count)
+        public
+        virtual
+        override
+        nonReentrant
+        returns (uint256[] memory tokenIds)
+    {
         requireExtension();
         tokenIds = new uint256[](count);
-        uint256 firstTokenId = _tokenCount+1;
+        uint256 firstTokenId = _tokenCount + 1;
         _tokenCount += count;
 
-        for (uint i; i < count;) {
-            tokenIds[i] = _mintExtension(to, "", 0, firstTokenId+i);
-            unchecked { ++i; }
+        for (uint256 i; i < count;) {
+            tokenIds[i] = _mintExtension(to, "", 0, firstTokenId + i);
+            unchecked {
+                ++i;
+            }
         }
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintExtensionBatch}.
      */
-    function mintExtensionBatch(address to, string[] calldata uris) public virtual override nonReentrant returns(uint256[] memory tokenIds) {
+    function mintExtensionBatch(address to, string[] calldata uris)
+        public
+        virtual
+        override
+        nonReentrant
+        returns (uint256[] memory tokenIds)
+    {
         requireExtension();
         tokenIds = new uint256[](uris.length);
-        uint256 firstTokenId = _tokenCount+1;
+        uint256 firstTokenId = _tokenCount + 1;
         _tokenCount += uris.length;
 
-        for (uint i; i < uris.length;) {
-            tokenIds[i] = _mintExtension(to, uris[i], 0, firstTokenId+i);
-            unchecked { ++i; }
+        for (uint256 i; i < uris.length;) {
+            tokenIds[i] = _mintExtension(to, uris[i], 0, firstTokenId + i);
+            unchecked {
+                ++i;
+            }
         }
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintExtensionBatch}.
      */
-    function mintExtensionBatch(address to, uint80[] calldata data) public virtual override nonReentrant returns(uint256[] memory tokenIds) {
+    function mintExtensionBatch(address to, uint80[] calldata data)
+        public
+        virtual
+        override
+        nonReentrant
+        returns (uint256[] memory tokenIds)
+    {
         requireExtension();
         tokenIds = new uint256[](data.length);
-        uint256 firstTokenId = _tokenCount+1;
+        uint256 firstTokenId = _tokenCount + 1;
         _tokenCount += data.length;
 
-        for (uint i; i < data.length;) {
-            tokenIds[i] = _mintExtension(to, "", data[i], firstTokenId+i);
-            unchecked { ++i; }
+        for (uint256 i; i < data.length;) {
+            tokenIds[i] = _mintExtension(to, "", data[i], firstTokenId + i);
+            unchecked {
+                ++i;
+            }
         }
     }
 
     /**
      * @dev Mint token via extension
      */
-    function _mintExtension(address to, string memory uri, uint80 data, uint256 tokenId) internal virtual returns(uint256) {
+    function _mintExtension(address to, string memory uri, uint80 data, uint256 tokenId)
+        internal
+        virtual
+        returns (uint256)
+    {
         if (tokenId == 0) {
             ++_tokenCount;
             tokenId = _tokenCount;
@@ -319,14 +386,22 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     /**
      * @dev See {ICreatorCore-setRoyalties}.
      */
-    function setRoyalties(address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyalties(address payable[] calldata receivers, uint256[] calldata basisPoints)
+        external
+        override
+        adminRequired
+    {
         _setRoyaltiesExtension(address(0), receivers, basisPoints);
     }
 
     /**
      * @dev See {ICreatorCore-setRoyalties}.
      */
-    function setRoyalties(uint256 tokenId, address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyalties(uint256 tokenId, address payable[] calldata receivers, uint256[] calldata basisPoints)
+        external
+        override
+        adminRequired
+    {
         require(_exists(tokenId), "Nonexistent token");
         _setRoyalties(tokenId, receivers, basisPoints);
     }
@@ -334,14 +409,24 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     /**
      * @dev See {ICreatorCore-setRoyaltiesExtension}.
      */
-    function setRoyaltiesExtension(address extension, address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyaltiesExtension(
+        address extension,
+        address payable[] calldata receivers,
+        uint256[] calldata basisPoints
+    ) external override adminRequired {
         _setRoyaltiesExtension(extension, receivers, basisPoints);
     }
 
     /**
      * @dev See {ICreatorCore-getRoyalties}.
      */
-    function getRoyalties(uint256 tokenId) external view virtual override returns (address payable[] memory, uint256[] memory) {
+    function getRoyalties(uint256 tokenId)
+        external
+        view
+        virtual
+        override
+        returns (address payable[] memory, uint256[] memory)
+    {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyalties(tokenId);
     }
@@ -349,7 +434,13 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     /**
      * @dev See {ICreatorCore-getFees}.
      */
-    function getFees(uint256 tokenId) external view virtual override returns (address payable[] memory, uint256[] memory) {
+    function getFees(uint256 tokenId)
+        external
+        view
+        virtual
+        override
+        returns (address payable[] memory, uint256[] memory)
+    {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyalties(tokenId);
     }
@@ -365,18 +456,18 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
     /**
      * @dev See {ICreatorCore-getFeeBps}.
      */
-    function getFeeBps(uint256 tokenId) external view virtual override returns (uint[] memory) {
+    function getFeeBps(uint256 tokenId) external view virtual override returns (uint256[] memory) {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyaltyBPS(tokenId);
     }
-    
+
     /**
      * @dev See {ICreatorCore-royaltyInfo}.
      */
     function royaltyInfo(uint256 tokenId, uint256 value) external view virtual override returns (address, uint256) {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyaltyInfo(tokenId, value);
-    } 
+    }
 
     /**
      * @dev See {IERC721Metadata-tokenURI}.
@@ -393,7 +484,7 @@ contract ERC721Creator is AdminControl, ERC721Base, ERC721CreatorCore {
         _setApproveTransferBase(extension);
     }
 
-    function _tokenExtension(uint256 tokenId) internal view override returns(address) {
+    function _tokenExtension(uint256 tokenId) internal view override returns (address) {
         uint16 extensionIndex = uint16(_tokenData[tokenId].data);
         if (extensionIndex == 0) return address(0);
         return _indexToExtension[extensionIndex];
