+++ contracts/ERC721CreatorImplementation.sol
@@ -26,8 +26,15 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     /**
      * @dev See {IERC165-supportsInterface}.
      */
-    function supportsInterface(bytes4 interfaceId) public view virtual override(ERC721Core, ERC721CreatorCore, AdminControlUpgradeable) returns (bool) {
-        return ERC721CreatorCore.supportsInterface(interfaceId) || ERC721Core.supportsInterface(interfaceId) || AdminControlUpgradeable.supportsInterface(interfaceId);
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(ERC721Core, ERC721CreatorCore, AdminControlUpgradeable)
+        returns (bool)
+    {
+        return ERC721CreatorCore.supportsInterface(interfaceId) || ERC721Core.supportsInterface(interfaceId)
+            || AdminControlUpgradeable.supportsInterface(interfaceId);
     }
 
     function _beforeTokenTransfer(address from, address to, uint256 tokenId, uint96 data) internal virtual override {
@@ -45,11 +52,14 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     /**
      * @dev See {ICreatorCore-registerExtension}.
      */
-    function registerExtension(address extension, string calldata baseURI, bool baseURIIdentical) external override adminRequired {
+    function registerExtension(address extension, string calldata baseURI, bool baseURIIdentical)
+        external
+        override
+        adminRequired
+    {
         requireNonBlacklist(extension);
         _registerExtension(extension, baseURI, baseURIIdentical);
     }
-
 
     /**
      * @dev See {ICreatorCore-unregisterExtension}.
@@ -103,9 +113,11 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     function setTokenURIExtension(uint256[] calldata tokenIds, string[] calldata uris) external override {
         requireExtension();
         require(tokenIds.length == uris.length, "Invalid input");
-        for (uint i; i < tokenIds.length;) {
+        for (uint256 i; i < tokenIds.length;) {
             _setTokenURIExtension(tokenIds[i], uris[i]);
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
     }
 
@@ -135,9 +147,11 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
      */
     function setTokenURI(uint256[] calldata tokenIds, string[] calldata uris) external override adminRequired {
         require(tokenIds.length == uris.length, "Invalid input");
-        for (uint i; i < tokenIds.length;) {
+        for (uint256 i; i < tokenIds.length;) {
             _setTokenURI(tokenIds[i], uris[i]);
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
     }
 
@@ -151,49 +165,74 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     /**
      * @dev See {IERC721CreatorCore-mintBase}.
      */
-    function mintBase(address to) public virtual override nonReentrant adminRequired returns(uint256) {
+    function mintBase(address to) public virtual override nonReentrant adminRequired returns (uint256) {
         return _mintBase(to, "", 0);
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintBase}.
      */
-    function mintBase(address to, string calldata uri) public virtual override nonReentrant adminRequired returns(uint256) {
+    function mintBase(address to, string calldata uri)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+        returns (uint256)
+    {
         return _mintBase(to, uri, 0);
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintBaseBatch}.
      */
-    function mintBaseBatch(address to, uint16 count) public virtual override nonReentrant adminRequired returns(uint256[] memory tokenIds) {
+    function mintBaseBatch(address to, uint16 count)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+        returns (uint256[] memory tokenIds)
+    {
         tokenIds = new uint256[](count);
-        uint256 firstTokenId = _tokenCount+1;
+        uint256 firstTokenId = _tokenCount + 1;
         _tokenCount += count;
 
-        for (uint i; i < count;) {
-            tokenIds[i] = _mintBase(to, "", firstTokenId+i);
-            unchecked { ++i; }
+        for (uint256 i; i < count;) {
+            tokenIds[i] = _mintBase(to, "", firstTokenId + i);
+            unchecked {
+                ++i;
+            }
         }
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintBaseBatch}.
      */
-    function mintBaseBatch(address to, string[] calldata uris) public virtual override nonReentrant adminRequired returns(uint256[] memory tokenIds) {
+    function mintBaseBatch(address to, string[] calldata uris)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+        returns (uint256[] memory tokenIds)
+    {
         tokenIds = new uint256[](uris.length);
-        uint256 firstTokenId = _tokenCount+1;
+        uint256 firstTokenId = _tokenCount + 1;
         _tokenCount += uris.length;
 
-        for (uint i; i < uris.length;) {
-            tokenIds[i] = _mintBase(to, uris[i], firstTokenId+i);
-            unchecked { ++i; }
+        for (uint256 i; i < uris.length;) {
+            tokenIds[i] = _mintBase(to, uris[i], firstTokenId + i);
+            unchecked {
+                ++i;
+            }
         }
     }
 
     /**
      * @dev Mint token with no extension
      */
-    function _mintBase(address to, string memory uri, uint256 tokenId) internal virtual returns(uint256) {
+    function _mintBase(address to, string memory uri, uint256 tokenId) internal virtual returns (uint256) {
         if (tokenId == 0) {
             ++_tokenCount;
             tokenId = _tokenCount;
@@ -214,7 +253,7 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     /**
      * @dev See {IERC721CreatorCore-mintExtension}.
      */
-    function mintExtension(address to) public virtual override nonReentrant returns(uint256) {
+    function mintExtension(address to) public virtual override nonReentrant returns (uint256) {
         requireExtension();
         return _mintExtension(to, "", 0, 0);
     }
@@ -222,7 +261,7 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     /**
      * @dev See {IERC721CreatorCore-mintExtension}.
      */
-    function mintExtension(address to, string calldata uri) public virtual override nonReentrant returns(uint256) {
+    function mintExtension(address to, string calldata uri) public virtual override nonReentrant returns (uint256) {
         requireExtension();
         return _mintExtension(to, uri, 0, 0);
     }
@@ -230,7 +269,7 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     /**
      * @dev See {IERC721CreatorCore-mintExtension}.
      */
-    function mintExtension(address to, uint80 data) public virtual override nonReentrant returns(uint256) {
+    function mintExtension(address to, uint80 data) public virtual override nonReentrant returns (uint256) {
         requireExtension();
         return _mintExtension(to, "", data, 0);
     }
@@ -238,52 +277,80 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     /**
      * @dev See {IERC721CreatorCore-mintExtensionBatch}.
      */
-    function mintExtensionBatch(address to, uint16 count) public virtual override nonReentrant returns(uint256[] memory tokenIds) {
+    function mintExtensionBatch(address to, uint16 count)
+        public
+        virtual
+        override
+        nonReentrant
+        returns (uint256[] memory tokenIds)
+    {
         requireExtension();
         tokenIds = new uint256[](count);
-        uint256 firstTokenId = _tokenCount+1;
+        uint256 firstTokenId = _tokenCount + 1;
         _tokenCount += count;
 
-        for (uint i; i < count;) {
-            tokenIds[i] = _mintExtension(to, "", 0, firstTokenId+i);
-            unchecked { ++i; }
+        for (uint256 i; i < count;) {
+            tokenIds[i] = _mintExtension(to, "", 0, firstTokenId + i);
+            unchecked {
+                ++i;
+            }
         }
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintExtensionBatch}.
      */
-    function mintExtensionBatch(address to, string[] calldata uris) public virtual override nonReentrant returns(uint256[] memory tokenIds) {
+    function mintExtensionBatch(address to, string[] calldata uris)
+        public
+        virtual
+        override
+        nonReentrant
+        returns (uint256[] memory tokenIds)
+    {
         requireExtension();
         tokenIds = new uint256[](uris.length);
-        uint256 firstTokenId = _tokenCount+1;
+        uint256 firstTokenId = _tokenCount + 1;
         _tokenCount += uris.length;
 
-        for (uint i; i < uris.length;) {
-            tokenIds[i] = _mintExtension(to, uris[i], 0, firstTokenId+i);
-            unchecked { ++i; }
+        for (uint256 i; i < uris.length;) {
+            tokenIds[i] = _mintExtension(to, uris[i], 0, firstTokenId + i);
+            unchecked {
+                ++i;
+            }
         }
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintExtensionBatch}.
      */
-    function mintExtensionBatch(address to, uint80[] calldata data) public virtual override nonReentrant returns(uint256[] memory tokenIds) {
+    function mintExtensionBatch(address to, uint80[] calldata data)
+        public
+        virtual
+        override
+        nonReentrant
+        returns (uint256[] memory tokenIds)
+    {
         requireExtension();
         tokenIds = new uint256[](data.length);
-        uint256 firstTokenId = _tokenCount+1;
+        uint256 firstTokenId = _tokenCount + 1;
         _tokenCount += data.length;
 
-        for (uint i; i < data.length;) {
-            tokenIds[i] = _mintExtension(to, "", data[i], firstTokenId+i);
-            unchecked { ++i; }
+        for (uint256 i; i < data.length;) {
+            tokenIds[i] = _mintExtension(to, "", data[i], firstTokenId + i);
+            unchecked {
+                ++i;
+            }
         }
     }
-    
+
     /**
      * @dev Mint token via extension
      */
-    function _mintExtension(address to, string memory uri, uint80 data, uint256 tokenId) internal virtual returns(uint256) {
+    function _mintExtension(address to, string memory uri, uint80 data, uint256 tokenId)
+        internal
+        virtual
+        returns (uint256)
+    {
         if (tokenId == 0) {
             ++_tokenCount;
             tokenId = _tokenCount;
@@ -326,14 +393,22 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     /**
      * @dev See {ICreatorCore-setRoyalties}.
      */
-    function setRoyalties(address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyalties(address payable[] calldata receivers, uint256[] calldata basisPoints)
+        external
+        override
+        adminRequired
+    {
         _setRoyaltiesExtension(address(0), receivers, basisPoints);
     }
 
     /**
      * @dev See {ICreatorCore-setRoyalties}.
      */
-    function setRoyalties(uint256 tokenId, address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyalties(uint256 tokenId, address payable[] calldata receivers, uint256[] calldata basisPoints)
+        external
+        override
+        adminRequired
+    {
         require(_exists(tokenId), "Nonexistent token");
         _setRoyalties(tokenId, receivers, basisPoints);
     }
@@ -341,14 +416,24 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     /**
      * @dev See {ICreatorCore-setRoyaltiesExtension}.
      */
-    function setRoyaltiesExtension(address extension, address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyaltiesExtension(
+        address extension,
+        address payable[] calldata receivers,
+        uint256[] calldata basisPoints
+    ) external override adminRequired {
         _setRoyaltiesExtension(extension, receivers, basisPoints);
     }
 
     /**
      * @dev See {ICreatorCore-getRoyalties}.
      */
-    function getRoyalties(uint256 tokenId) external view virtual override returns (address payable[] memory, uint256[] memory) {
+    function getRoyalties(uint256 tokenId)
+        external
+        view
+        virtual
+        override
+        returns (address payable[] memory, uint256[] memory)
+    {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyalties(tokenId);
     }
@@ -356,7 +441,13 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     /**
      * @dev See {ICreatorCore-getFees}.
      */
-    function getFees(uint256 tokenId) external view virtual override returns (address payable[] memory, uint256[] memory) {
+    function getFees(uint256 tokenId)
+        external
+        view
+        virtual
+        override
+        returns (address payable[] memory, uint256[] memory)
+    {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyalties(tokenId);
     }
@@ -372,18 +463,18 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
     /**
      * @dev See {ICreatorCore-getFeeBps}.
      */
-    function getFeeBps(uint256 tokenId) external view virtual override returns (uint[] memory) {
+    function getFeeBps(uint256 tokenId) external view virtual override returns (uint256[] memory) {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyaltyBPS(tokenId);
     }
-    
+
     /**
      * @dev See {ICreatorCore-royaltyInfo}.
      */
     function royaltyInfo(uint256 tokenId, uint256 value) external view virtual override returns (address, uint256) {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyaltyInfo(tokenId, value);
-    } 
+    }
 
     /**
      * @dev See {IERC721Metadata-tokenURI}.
@@ -400,7 +491,7 @@ contract ERC721CreatorImplementation is AdminControlUpgradeable, ERC721Upgradeab
         _setApproveTransferBase(extension);
     }
 
-    function _tokenExtension(uint256 tokenId) internal view override returns(address) {
+    function _tokenExtension(uint256 tokenId) internal view override returns (address) {
         uint16 extensionIndex = uint16(_tokenData[tokenId].data);
         if (extensionIndex == 0) return address(0);
         return _indexToExtension[extensionIndex];
