+++ contracts/extensions/ERC1155/ERC1155CreatorExtensionBurnable.sol
@@ -16,30 +16,40 @@ import "./IERC1155CreatorExtensionBurnable.sol";
  * the creator of a token created.
  */
 abstract contract ERC1155CreatorExtensionBurnable is AdminControl, IERC1155CreatorExtensionBurnable {
-
-    mapping (uint256 => address) private _tokenCreators;
+    mapping(uint256 => address) private _tokenCreators;
 
     /**
      * @dev See {IERC165-supportsInterface}.
      */
     function supportsInterface(bytes4 interfaceId) public view virtual override(AdminControl, IERC165) returns (bool) {
-        return interfaceId == type(IERC1155CreatorExtensionBurnable).interfaceId
-            || super.supportsInterface(interfaceId);
+        return interfaceId == type(IERC1155CreatorExtensionBurnable).interfaceId || super.supportsInterface(interfaceId);
     }
 
     /**
      * @dev batch mint a token
      */
-    function mintNew(address creator, address[] calldata to, uint256[] calldata amounts, string[] calldata uris) external adminRequired returns (uint256[] memory) {
+    function mintNew(address creator, address[] calldata to, uint256[] calldata amounts, string[] calldata uris)
+        external
+        adminRequired
+        returns (uint256[] memory)
+    {
         return _mintNew(creator, to, amounts, uris);
     }
 
-    function _mintNew(address creator, address[] calldata to, uint256[] calldata amounts, string[] calldata uris) internal returns (uint256[] memory tokenIds) {
-        require(ERC165Checker.supportsInterface(creator, type(IERC1155CreatorCore).interfaceId), "creator must implement IERC1155CreatorCore");
+    function _mintNew(address creator, address[] calldata to, uint256[] calldata amounts, string[] calldata uris)
+        internal
+        returns (uint256[] memory tokenIds)
+    {
+        require(
+            ERC165Checker.supportsInterface(creator, type(IERC1155CreatorCore).interfaceId),
+            "creator must implement IERC1155CreatorCore"
+        );
         tokenIds = IERC1155CreatorCore(creator).mintExtensionNew(to, amounts, uris);
-        for (uint i; i < tokenIds.length;) {
+        for (uint256 i; i < tokenIds.length;) {
             _tokenCreators[tokenIds[i]] = creator;
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
     }
 
@@ -47,11 +57,11 @@ abstract contract ERC1155CreatorExtensionBurnable is AdminControl, IERC1155Creat
      * @dev See {IERC1155CreatorExtensionBurnable-onBurn}.
      */
     function onBurn(address, uint256[] calldata tokenIds, uint256[] calldata) public virtual override {
-        for (uint i; i < tokenIds.length;) {
+        for (uint256 i; i < tokenIds.length;) {
             require(_tokenCreators[tokenIds[i]] == msg.sender, "Can only be called by token creator");
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
     }
-
-
 }
